Move integration tests into tests directory
